requirements.txt
Flask
flask-cors
tensorflow==2.11.0
tensorflow-hub
Pillow
numpy
